move wgt avg function to utilities
